feat: HTML preview of the search results page
It was already possible to preview pages coming from disk files and
`resource://` URIs, but not the search results page. This commit extends
HTML preview support to search results.

Function `preview_clicked` is extended to use the viewer instance's
`backing_file` instead of the `current_file` when the latter is a
`search://` URI.

Markdown links in the search results page are modified so that they use
relative paths. This makes the exported HTML page easier to repurpose
following some of the examples from the MDVIEW MTX wiki.