fix: empty link text in search results page
Recall that the search results page displays the first heading found
in the file as the link text of a document matching the search terms.

Some Markdown authors use empty ATX headings to insert vertical space.
For instance, the first heading of the Fatdog64 903 FAQ files is
"# \u00A0", and consequently before this commit, the search results page
displayed empty link text as the titles of all documents. This commit
tries harder to find a non-empty representation for the link text.

First, now the regex matcher for Markdown ATX headings in function
`mtx_cmm_regex_astx` skips "empty" ATX headings -- where emptiness
means comprising only horizontal or vertical Unicode separator and
white space code points).

Changing the regex matcher for SETEXT headings is too complicated. For
those, function `_file_search` puts in place a different fix. All runs
of Unicode separators are folded to an ASCII blank. If the resulting
string does not contain a non-blank character it is discarded, and the
file name is used as the document title.